[SIDE-1196] Refactor logo components
