+++ docs/src/components/aptitudes/shared-links.ts
@@ -1,0 +1,3 @@
+export interface ILogoProps {
+  fillColor?: string;
+}
